microservicio-4/requirements.txt
django==4.1.6
django_rest_framework==0.1.0
django-cors-headers==4.3.1
tensorflow==2.14.0